% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_toplevel}
\alias{count_toplevel}
\title{Determines which of the dependencies are directly
referenced in the user's code}
\usage{
count_toplevel()
}
\description{
Determines which of the dependencies are directly
referenced in the user's code
}
