% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{countdeps}
\alias{countdeps}
\title{Count Dependencies}
\usage{
countdeps(dir = ".", keepLockfile = FALSE)
}
\arguments{
\item{dir}{The project directory containing the code you'd like to scan for dependencies}

\item{keepLockfile}{Whether or not to create a packrat lock file in the directory with the identified dependencies}
}
\description{
Count Dependencies
}
